dom/worklet/tests/test_dump.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worklet - Console</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">

function configureTest() {
  return SpecialPowers.pushPrefEnv(
    {"set": [["dom.audioWorklet.enabled", true],
             ["dom.worklet.enabled", true]]});
}

// This function is called into an iframe.
function runTestInIframe() {
  audioWorklet.import("worklet_dump.js")
  .then(() => {
    ok(true, "All good!");
    SimpleTest.finish();
  });
}
</script>
</body>
</html>
